refactor(tab): rm .js from imports
