+++ packages/tab/src/react/index.tsx
@@ -1,6 +1,6 @@
-import List from './list.js'
-import ListItem from './list-item.js'
-import Panel from './panel.js'
+import List from './list'
+import ListItem from './list-item'
+import Panel from './panel'
 
 export default {
   List,
